moved useful functions from EnergyCarrier to Misc
