+++ src/Hive.IO/Hive.IO/Misc.cs
@@ -14,5 +14,60 @@ namespace Hive.IO
         public static readonly int[] DaysPerMonth = new int[12] { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
         public const int HoursPerDay = 24;
         public const int HoursPerYear = 8760;
+
+        public static double[] GetAverageMonthlyValue(double[] annualTimeSeries)
+        {
+
+            int months = 12;
+            double[] monthlyTimeSeries = new double[months];
+            int sumOfDays = 0;
+            for (int t = 0; t < months; t++)
+            {
+                int startIndex = sumOfDays * Misc.HoursPerDay;
+                int daysThisMonth = Misc.DaysPerMonth[t];
+                sumOfDays += daysThisMonth;
+                int endIndex = sumOfDays * Misc.HoursPerDay;
+                double average = 0.0;
+                for (int i = startIndex; i < endIndex; i++)
+                {
+                    double temp = annualTimeSeries[i];
+                    if (double.IsNaN(temp))
+                        temp = 0.0;
+                    average += temp;
+                }
+
+                average /= (daysThisMonth * Misc.HoursPerDay);
+                //double average = Enumerable.Range(startIndex, endIndex).Select(i => annualTimeSeries[i]).Average();
+                monthlyTimeSeries[t] = average;
+            }
+
+            return monthlyTimeSeries;
+        }
+
+
+        public static double[] GetCumulativeMonthlyValue(double[] annualTimeSeries)
+        {
+            int months = 12;
+            double[] monthlyTimeSeries = new double[months];
+            int sumOfDays = 0;
+            for (int t = 0; t < months; t++)
+            {
+                int startIndex = sumOfDays * Misc.HoursPerDay;
+                int daysThisMonth = Misc.DaysPerMonth[t];
+                sumOfDays += daysThisMonth;
+                int endIndex = sumOfDays * Misc.HoursPerDay;
+                double sum = 0.0;
+                for (int i = startIndex; i < endIndex; i++)
+                {
+                    double temp = annualTimeSeries[i];
+                    if (double.IsNaN(temp))
+                        temp = 0.0;
+                    sum += temp;
+                }
+                monthlyTimeSeries[t] = sum;
+            }
+
+            return monthlyTimeSeries;
+        }
     }
 }
